+++ engine/apps/schedules/tests/test_on_call_schedule.py
@@ -1158,6 +1158,49 @@ def test_schedule_related_users(make_organization, make_user_for_organization, m
     assert set(users) == set([user_a, user_d, user_e])
 
 
+@pytest.mark.django_db
+def test_schedule_related_users_usernames(
+    make_organization, make_user_for_organization, make_on_call_shift, make_schedule
+):
+    """
+    Check different usernames, including those with special characters and uppercase letters
+    """
+    organization = make_organization()
+    schedule = make_schedule(
+        organization,
+        schedule_class=OnCallScheduleWeb,
+        name="test_web_schedule",
+    )
+
+    now = timezone.now().replace(hour=0, minute=0, second=0, microsecond=0)
+    start_date = now - timezone.timedelta(days=7)
+
+    # Check different usernames, including those with special characters and uppercase letters
+    usernames = ["test", "test.test", "test.test@test.test", "TEST.TEST@TEST.TEST"]
+    users = [make_user_for_organization(organization, username=u) for u in usernames]
+    # clear users pks <-> organization cache (persisting between tests)
+    memoized_users_in_ical.cache_clear()
+
+    for user in users:
+        data = {
+            "start": start_date,
+            "rotation_start": start_date,
+            "duration": timezone.timedelta(hours=1),
+            "priority_level": 1,
+            "frequency": CustomOnCallShift.FREQUENCY_DAILY,
+            "schedule": schedule,
+        }
+        on_call_shift = make_on_call_shift(
+            organization=organization, shift_type=CustomOnCallShift.TYPE_ROLLING_USERS_EVENT, **data
+        )
+        on_call_shift.add_rolling_users([[user]])
+
+    schedule.refresh_ical_file()
+    schedule.refresh_from_db()
+
+    assert set(schedule.related_users()) == set(users)
+
+
 @pytest.mark.django_db(transaction=True)
 def test_filter_events_none_cache_unchanged(
     make_organization, make_user_for_organization, make_schedule, make_on_call_shift
@@ -2329,11 +2372,16 @@ def test_swap_request_no_changes(
 
     # setup swap requests
     tomorrow = today + timezone.timedelta(days=1)
+    # user not in schedule
     make_shift_swap_request(schedule, other_user, swap_start=today, swap_end=tomorrow)
+    # deleted request
     make_shift_swap_request(schedule, user, swap_start=today, swap_end=tomorrow, deleted_at=today)
+    # swap request in the past
     make_shift_swap_request(
         schedule, user, swap_start=today - timezone.timedelta(days=7), swap_end=tomorrow - timezone.timedelta(days=7)
     )
+    # untaken swap in progress (past due)
+    make_shift_swap_request(schedule, user, swap_start=today - timezone.timedelta(days=1), swap_end=tomorrow)
 
     events_after = schedule.filter_events(today, today + timezone.timedelta(days=2))
     assert events_before == events_after