+++ CHANGELOG.md
@@ -7,10 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Add mobile app push notifications for shift swap requests by @vadimkerr ([#2717](https://github.com/grafana/oncall/pull/2717))
+
+### Changed
+
+- Skip past due swap requests when calculating events ([2718](https://github.com/grafana/oncall/pull/2718))
+
 ### Fixed
 
 - Fix schedule final_events datetime filtering when splitting override ([#2715](https://github.com/grafana/oncall/pull/2715))
 - Fix swap requests event filter limits in schedule events ([#2716](https://github.com/grafana/oncall/pull/2716))
+- Fix Alerting contact point auto-creation ([2721](https://github.com/grafana/oncall/pull/2721))
 
 ## v1.3.21 (2023-08-01)